requirements.txt
watchdog
pywin32
ijson
aiofiles
python-dotenv
asyncpg
tenacity
pytest
